requirements.txt
Flask
tensorflow
numpy
Pillow
python-dotenv
google-generativeai
requests
gunicorn